Overian_cancer_prediction/requirements.txt
flask
pandas
scikit-learn
xgboost
transformers
torch
accelerate
python-dotenv
requests
